+++ nuxeo-drive-client/tests/test_local_creations.py
@@ -134,3 +134,25 @@ class TestLocalCreations(UnitTestCase):
         children = local.get_children_info('/')
         self.assertEqual(len(children), 1)
         self.assertEqual(children[0].name, filename_upper)
+
+    def test_local_create_folders_with_dots(self):
+        """ Check that folders containing dots are well synced. """
+
+        remote = self.remote_document_client_1
+        local = self.local_client_1
+        engine = self.engine_1
+
+        engine.start()
+        self.wait_sync(wait_for_async=True)
+
+        folder1 = 'Affaire.1487689320370'
+        folder2 = 'Affaire.1487689320.370'
+        local.make_folder('/', folder1)
+        local.make_folder('/', folder2)
+        self.wait_sync()
+
+        # Check
+        assert remote.exists('/' + folder1)
+        assert remote.exists('/' + folder2)
+        assert local.exists('/' + folder1)
+        assert local.exists('/' + folder2)